==> 8-item_item_build.ipynb <==
import pandas as pd

df_rate = pd.read_csv('../data/play_rate.csv')

df = pd.read_csv('../data/play_ds.csv')

df.head(5)

def cleanPlayTime(row):
    if row['play_time'] == 0:
        return 0
    if '>' in row['play_time'] :
        # print s.split('>')
        return row['play_time'].split('>')[0]
    else:
        return row['play_time']

df['play_time'].fillna('0', inplace=True)

df['play_time_2'] = df.apply(cleanPlayTime,axis = 1)

def calculateTime(row):
    if int(row['song_length']) == 0:
        #print("processed")
        return 0
    else:
        #print("processed")
        return float(row['play_time_2']) / float(row['song_length'])

df[['play_time_2']] = df[['play_time_2']].astype(float)
df['play_time_2'].fillna('0', inplace=True)
df['song_length'].fillna('0', inplace=True)

df.iloc[[748]]

df['play_ratio'] = df.apply(calculateTime,axis = 1)

df.head(5)

df.to_csv('../data/play_ds_extention.csv')

df_rate = df.groupby(['uid','song_id'])
df_rate = df.drop(['device','date','play_time','song_length','play_time_2'], axis=1)
df_rate = df_rate.groupby(['uid', 'song_id']).agg({'play_ratio':'sum'})
df_rate = df_rate.reset_index()
df_rate.to_csv('../data/play_rate.csv')

df_rate.head(5)

df_rate[df_rate['play_ratio'] >= 1].count()

df_rate[df_rate['play_ratio'] < 1].count()

# Due to we have too much data, we have to resample based on play_ratio
from sklearn.utils import resample
# Separate majority and minority classes
df_majority = df_rate[df_rate.play_ratio < 1]
df_minority = df_rate[df_rate.play_ratio >= 1]
 
# Downsample majority class
df_majority_downsampled = resample(df_majority, 
                                 replace=False,    # sample without replacement
                                 n_samples= ,     # to match minority class
                                 random_state=123) # reproducible results
 
# Combine minority class with downsampled majority class
df_downsampled = pd.concat([df_majority_downsampled, df_minority])
 
# Display new class counts
df_downsampled.balance.value_counts()
# 1    49
# 0    49
# Name: balance, dtype: int64

df_rate = df_rate.dropna()

df_rate.to_csv('../data/play_rate.csv')

df_rate = pd.read_csv('../data/play_rate.csv')

df_rate.head(5)

df_rate = df_rate[df_rate['song_id'] > 0]

df_rate.size

df_rate['freq'] = df_rate.groupby('song_id')['song_id'].transform('count')

df_rate = df_rate[df_rate['freq'] >= 10]

df_rate.size

len(df_rate.song_id.unique())

from scipy import sparse
from sklearn.metrics.pairwise import cosine_similarity

# convert uid to a smaller list
user_id_lst = df_rate['uid'].unique() 
user_id_dict = {k: v for v, k in enumerate(user_id_lst)}

# convert song id to a smaller list
song_id_lst = df_rate['song_id'].unique() 
song_id_dict = {k: v for v, k in enumerate(song_id_lst)}

df_rate['user_id_map'] = df_rate.apply(lambda x: user_id_dict[x['uid']],axis = 1)
df_rate['song_id_map'] = df_rate.apply(lambda x: song_id_dict[x['song_id']],axis = 1) 

highest_user_id = df_rate.user_id_map.max() + 1
highest_song_id = df_rate.song_id_map.max() + 1
ratings_mat = sparse.lil_matrix((highest_user_id, highest_song_id)) 
ratings_mat

from math import fabs
for _, row in df_rate.iterrows():
# Due to some data error, use absolute to correct negative play time
# subtract 1 from id's due to match 0 indexing 
    ratings_mat[row.user_id_map-1, row.song_id_map-1] = fabs(row.play_ratio)

# # item_item recommendation
# utility_mat = ratings_mat
# Item-Item Similarity Matrix
item_sim_mat = cosine_similarity(ratings_mat.T)

import numpy as np
least_to_most_sim_indexes = np.argsort(item_sim_mat, axis=1)
# Neighborhoods
neighborhood_size = 75
neighborhoods = least_to_most_sim_indexes[:, -neighborhood_size:]

neighborhoods.shape

from time import time
user_id = 12333

n_users = ratings_mat.shape[0] 
n_items = ratings_mat.shape[1]

start_time = time()
items_rated_by_this_user = ratings_mat[user_id].nonzero()[1] 
# Just initializing so we have somewhere to put rating preds 
out = np.zeros(n_items)
for item_to_rate in range(n_items):
    relevant_items = np.intersect1d(neighborhoods[item_to_rate], 
                                    items_rated_by_this_user,
                                    assume_unique=True) 
    # assume_unique speeds up intersection 
    out[item_to_rate] = ratings_mat[user_id, relevant_items] * item_sim_mat[item_to_rate, relevant_items] / item_sim_mat[item_to_rate, relevant_items].sum()

pred_ratings = np.nan_to_num(out)
print(pred_ratings)
print("Execution time: %f seconds" % (time()-start_time))

# Get final recommendation # Recommend n movies 
n = 10

# Get item indexes sorted by predicted rating 
item_index_sorted_by_pred_rating = list(np.argsort(pred_ratings))[::-1]

# Find items that have been rated by user 
items_rated_by_this_user = ratings_mat[user_id].nonzero()[1]

# We want to exclude the items that have been rated by user 
unrated_items_by_pred_rating = [item for item in item_index_sorted_by_pred_rating if item not in items_rated_by_this_user]

unrated_items_by_pred_rating[:n]

# Get song information
df_song_info = pd.read_csv('../data/play/song_info.csv')

df_song_info.head(5)

df_song_info = df_song_info.drop(['Unnamed: 0'],axis=1)

unrated_items_by_pred_rating[0]

df_rate[df_rate['song_id_map'] == 1]['song_id'].unique()[0]

df_rate.head(5)

# Show the corresponding 
recommend_list = []
for i in range(10):
    song_map_id = unrated_items_by_pred_rating[i]
    song_id = df_rate[df_rate['song_id_map'] == song_map_id]['song_id'].unique()[0]
    recommend_list.append(df_song_info[df_song_info['song_id'] == song_id].iloc[0])
recommend_list

df_rate.to_csv('../data/play_rate.csv')
