==> 9-NMF.ipynb <==
import pandas as pd

df_rate = pd.read_csv('../data/play_rate.csv')

df_rate.head(5)

condition1 = df_rate['play_ratio'] >= 0
condition2 = df_rate['freq'] >= 10
df_rate = df_rate[condition1 & condition2]

df_rate.size

len(df_rate.song_id.unique())

from scipy import sparse
highest_user_id = df_rate.user_id_map.max() 
highest_song_id = df_rate.song_id_map.max() 
ratings_mat = sparse.lil_matrix((highest_user_id, highest_song_id)) 
ratings_mat

for _, row in df_rate.iterrows():
    # subtract 1 from id's due to match 0 indexing 
    ratings_mat[row.user_id_map-1, row.song_id_map-1] = row.play_ratio

from sklearn.decomposition import TruncatedSVD

def fit_uvd(M,k):
    # use TruncatedSVD to realize UVD 
    svd = TruncatedSVD(n_components=k, n_iter=7, random_state=0) 
    svd.fit(M)

    V = svd.components_ 
    U = svd.transform(M)
    
    return U,V, svd

U,V,svd = fit_uvd(ratings_mat,200)

print(U.shape,V.shape)

import numpy as np
# reconstruct 
ratings_mat_fitted = U.dot(V) # U*V
errs = np.array((ratings_mat-ratings_mat_fitted).flatten()).squeeze() 
mask = np.array((ratings_mat.todense()).flatten()).squeeze()>0

# calculate errs
mse = np.mean(errs[mask]**2) 
average_abs_err = abs(errs[mask]).mean() 
print(mse) 
print(average_abs_err)

# compare with another way to reconstruct matrix # with the above "tranformed to the new space and back" language # without the UV language, we can do:

# reconstruct M 
with inverse_transform 
ratings_mat_fitted_2 = svd.inverse_transform(svd.transform(ratings_mat)) 
ratings_mat_fitted = U.dot(V) 
print(sum(sum(ratings_mat_fitted - ratings_mat_fitted_2)))

user_id = 12311
n = 10

pred_ratings = ratings_mat_fitted[user_id,:] 
item_index_sorted_by_pred_rating = list(np.argsort(pred_ratings))[::-1]

items_rated_by_this_user = ratings_mat[user_id].nonzero()[1]

unrated_items_by_pred_rating = [item for item in item_index_sorted_by_pred_rating if item not in items_rated_by_this_user]

unrated_items_by_pred_rating[:n]

df_song_info = pd.read_csv('../data/play/song_info.csv')
df_song_info = df_song_info.drop(['Unnamed: 0'],axis=1)

recommend_list = []
for i in range(10):
    song_map_id = unrated_items_by_pred_rating[i]
    song_id = df_rate[df_rate['song_id_map'] == song_map_id]['song_id'].unique()[0]
    recommend_list.append(df_song_info[df_song_info['song_id'] == song_id].iloc[0])
recommend_list
